# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Trabajo",
        "description": "Documentación de la API desarrollada en Laravel para la bolsa de trabajo.",
        "contact": {
            "name": "Soporte",
            "email": "soporte@tuapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor Laravel local"
        }
    ],
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener todos los roles",
                "operationId": "9fa07677d6ca82e42fb3002f1a229295",
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "roles": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo rol",
                "operationId": "a3549828f62dfa78f6e9a3074c416372",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rol": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rol creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener un rol específico",
                "operationId": "c6c2d9200d74017052e538857039d540",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rol": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Actualizar un rol existente",
                "operationId": "d2f7da4cc75a13746482ef71d7521150",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Administrador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rol": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rol actualizado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un rol",
                "operationId": "429d91b2d5c8a8aa91002cabca373303",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol eliminado exitosamente"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            }
        },
...
    },
...
}